Move example to example git repo
