--- example/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "test",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "Glarce": "../",
-    "cross-env": "^5.2.0"
-  },
-  "scripts": {
-    "build": "cross-env production=true node app",
-    "serve": "node app",
-    "test": "rm -rf node_modules/Glarce && yarn add ../ && yarn serve"
-  }
-}
