--- public/videos
@@ -1,1 +1,0 @@
-../../../media/videos
