--- src/app.json
@@ -1,1 +1,0 @@
-[{"scanType":"barcode","scan":5,"contentType":"video","videoData":{"id":1,"url":"videos/spook.webm","extension":"webm","interactive":[{"type":"alert","title":"Pause","text":"Continue watching...","sec":1,"executed":false,"id":10},{"type":"question","title":"Jump","text":"Do you want to jump to 5 seconds","buttons":[{"text":"ok","jumpTo":5},{"text":"no"}],"sec":2,"executed":false,"id":11},{"type":"end","sec":3,"executed":false,"id":12}],"height":3,"width":5.333333333333333}}]
